backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
openai==1.57.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.0
pydantic-settings==2.6.1
pydantic_core==2.18.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.24.0
